drf-server/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-v5==1.0.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==2.1.1
JustWatch==0.5.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
zipp==3.4.1
